2009-01-22  Alan McGovern  <amcgovern@novell.com>
        * KnownTypeCollection.cs : The byte overload doesn't exist in
        Silverlight, so cast the byte to an int instead.



svn path=/trunk/mcs/; revision=124205

+++ mcs/class/System.Runtime.Serialization/System.Runtime.Serialization/KnownTypeCollection.cs
@@ -224,7 +224,7 @@ namespace System.Runtime.Serialization
 			case TypeCode.Boolean:
 				return XmlConvert.ToString ((bool) obj);
 			case TypeCode.Byte:
-				return XmlConvert.ToString ((byte) obj);
+				return XmlConvert.ToString ((int)((byte) obj));
 			case TypeCode.Char:
 				return XmlConvert.ToString ((uint) (char) obj);
 			case TypeCode.DateTime:
